LPD-44396 Refactor BaseCaptchaTestCase into a util
